Make the stash manager a class property
The current setup assumed that we had one Stash instance per process
and so if the lock property on the class was initialised then we must
also have initialised the manager. The way that websockets handlers
are loaded breaks this assumption, so we need to ensure that the
lifecycle of the lock and of the manager object are identical. Sharing
the manager state between instances makes sense since we're already
sharing the proxy we get from the stash.