[GridNG] Not crashing when we have grid with a subgridded axis
This CL fixes a crash that happens when we try to position an item in a
grid with a subgridded axis and an auto repeater track list.

According to the spec https://www.w3.org/TR/css-grid-2/#subgrid-listing
"If there is no parent grid, this value is equivalent to the initial
value, none." In |grid_position_resolver| we check for this and only use
the serialized grid line names when they turn out to be valid.

Added two new tests to validate the behavior with in-flow and out-of-
flow items.

Bug: 618969, 1275398
Change-Id: I2bc2e368b73329cd7f01af9d3849a61e869f5db4