[FlexNG] Simple multi-line row flex fragmentation
Add support for simple multi-line row flex fragmentation. In order to
get this case working, we needed to update when we consider an item
to have container separation in the case of rows (i.e. when the item's
row is not the first in a given fragmentainer).

There was also an issue in the NGFlexItemIterator when handling
break tokens in the multi-line case. More specifically, we won't know
the current item's index or line index ahead of time if we're
processing a break token. Move the calculation of |current_item_idx| and
|current_line_idx| to accommodate this case.

Bug: 660611
Change-Id: I67d5c59bd541cc42b22bb462ef840d55ce5ad67e